refactor :: 코드 분해
